Packaging: moved all files imported from main into submodules
