+++ core/GlobalRules.py
