+++ core/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Fri May 10 13:26:49 2024
+
+@author: zetadin
+"""
+
