+++ entities/resources/Resources.py
@@ -8,7 +8,7 @@ Created on Thu May 16 15:02:29 2024
 
 import numpy as np
 import numpy.typing as npt
-import GlobalRules
+import core.GlobalRules as GlobalRules
 from entities.Entity import SpriteEntity
 from utils.AssetFactory import assetFactory
 from utils.Geometry import expandPeriodicImages
