+++ requirements.txt
@@ -3,3 +3,4 @@ numpy>=1.24.1
 pybind11==2.11.1
 pytest==8.2.0
 pyinstaller==6.9.0
+tqdm==4.66.4
