+++ wavelets/Wavelet.py
@@ -8,7 +8,7 @@ Created on Fri May 10 13:41:03 2024
 
 import numpy as np
 import numpy.typing as npt
-import GlobalRules
+import core.GlobalRules as GlobalRules
 from utils.Geometry import expandPeriodicImages
 
 class Wavelet:
